generate

+++ src/main/java/com/github/truongbb/geneticschoolbusroutingalgorithm/config/SchoolBusConfiguration.java
@@ -3,6 +3,7 @@ package com.github.truongbb.geneticschoolbusroutingalgorithm.config;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.relational.core.sql.In;
 
 @Getter
 @Configuration
@@ -21,7 +22,7 @@ public class SchoolBusConfiguration {
     Integer elitisNumber;
 
     @Value("${application.population-size}")
-    Long populationSize;
+    Integer populationSize;
 
     @Value("${application.generation-number}")
     Integer generationNumber;
